chore: testing fetch events from undeployed deposit contracts
