+++ tests/utils.py
@@ -1,4 +1,7 @@
 from brownie import web3, chain, history
+from web3._utils.filters import construct_event_filter_params
+from web3._utils.events import get_event_data
+import json
 
 
 def cleanHexStr(thing):
@@ -66,10 +69,6 @@ def trimToShortest(lists):
     return minLen
 
 
-def null_sig(nonce):
-    return (0, 0, nonce)
-
-
 def getValidTranIdxs(tokens, amounts, prevBal, tok):
     # Need to know which index that native transfers start to fail since they won't revert the tx, but won't send the expected amount
     cumulEthTran = 0
@@ -116,3 +115,64 @@ def calculateGasTransaction(txReceipt):
     base_fee = web3.eth.get_block(txReceipt.block_number).baseFeePerGas
     priority_fee = txReceipt.gas_price - base_fee
     return (txReceipt.gas_used * base_fee) + (txReceipt.gas_used * priority_fee)
+
+
+def get_contract_object(path_to_contract, address):
+    ## path_to_contract from contracts folder. If a contract under the contracts folder, just the name of the contract.
+    with open("build/contracts/" + path_to_contract + ".json") as f:
+        info_json = json.load(f)
+    abi = info_json["abi"]
+    return web3.eth.contract(address=address, abi=abi)
+
+
+# In order to get the event from a contract do "get_contract_object("contract_name", contract_address).events.event_name
+def fetch_events(
+    event,
+    argument_filters=None,
+    from_block=None,
+    to_block="latest",
+    address=None,
+    topics=None,
+):
+    """Get events using eth_getLogs API.
+
+    This is a stateless method, as opposite to createFilter and works with
+    stateless nodes like QuikNode and Infura.
+
+    :param event: Event instance from your contract.events
+    :param argument_filters:
+    :param from_block: Start block. Use 0 for all history/
+    :param to_block: Fetch events until this contract
+    :param address:
+    :param topics:
+    :return:
+    """
+
+    if from_block is None:
+        raise TypeError("Missing mandatory keyword argument to getLogs: from_Block")
+
+    abi = event._get_event_abi()
+    abi_codec = event.web3.codec
+
+    # Set up any indexed event filters if needed
+    argument_filters = dict()
+    _filters = dict(**argument_filters)
+
+    data_filter_set, event_filter_params = construct_event_filter_params(
+        abi,
+        abi_codec,
+        contract_address=event.address,
+        argument_filters=_filters,
+        fromBlock=from_block,
+        toBlock=to_block,
+        address=address,
+        topics=topics,
+    )
+
+    # Call node over JSON-RPC API
+    logs = event.web3.eth.get_logs(event_filter_params)
+
+    # Convert raw binary event data to easily manipulable Python objects
+    for entry in logs:
+        data = get_event_data(abi_codec, abi, entry)
+        yield data
